CNS-140: dockerize lavad setup
Update the Dockerfile to generate docker image that can run lava node.
(Implemented using the artifacts in docker/ directory.)

To start a lava node in a container, run the command:
    docker-compose --profile node --env-file env -f docker/docker-compose.yml up

Edit the file `env` to adjust the (external) ports used by this node (see that
file for the defaults).

It will execute the script `start_node.sh` (from docker/), which follows
the instructions for starting a node (with cosmovisor).