CNS-140: dockerize lavad setup: portal
Update the Dockerfile to generate docker image that can run lava portal.
(Implemented using the artifacts in docker/ directory.)

To start a lava portal in a container, run the command:
    docker-compose --profile portal --env-file env -f docker/docker-compose.yml up

Edit the file `env` to adjust the portal parameters used by this portal (see that
file for the defaults).

It will execute the script `start_portal.sh` (from docker/), which follows
the instructions for starting a portal. If needed, it will also do stake-client
(including if the stake amount is to be raised).