CNS-140: dockerize lavad setup: provider
Update the Dockerfile to generate docker image that can run lava provider.
(Implemented using the artifacts in docker/ directory.)

To start a lava provider in a container, run the command:
    docker-compose --profile provider \
        --env-file docker/env -f docker/docker-compose.yml up

Edit the file `env` to adjust the portal parameters used by this provider (see
that file for the defaults).

It will execute the script `start_provider.sh` (from docker/), which follows
the instructions for starting a provider. If needed, it will also execute the
stake-provider (including if the stake amount is to be raised).