CNS-250: introduce spec inheritance syntax
Extend the spec syntax to add the ability for a spec to inherit/import another
spec. With spec inheritance, changes in an ancestor spec (a.k.a base spec)
would immediately propagate to all descendants specs that inherit from it.

Spec inheritance is recursive, and applied only the APIs defined (not general
spec parameters). APIs in current spec override inherited APIs, but duplicates
(from multiple ancestors) are not allowed. An override with 'Enabled: false'
effectively removes an API). Spec inheritance is valitaded when submitting a
spec proposal.