CNS 250: Introduce Spec inheritance syntax (#288)
* CNS-250: introduce spec inheritance syntax

Extend the spec syntax to add the ability for a spec to inherit/import another
spec. With spec inheritance, changes in an ancestor spec (a.k.a base spec)
would immediately propagate to all descendants specs that inherit from it.

Spec inheritance is recursive, and applied only the APIs defined (not general
spec parameters). APIs in current spec override inherited APIs, but duplicates
(from multiple ancestors) are not allowed. An override with 'Enabled: false'
effectively removes an API). Spec inheritance is valitaded when submitting a
spec proposal.

* CNS-250: query spec to return full (expanded) spec

For compatibility, query for spec(s) should return fully expanded spec(s).
This is also serves stake-provider when they query the spec, as they need
to validate and implement inherited APIs too.

* CNS-250: allow duplicate imported api with override

Allow import with duplicate API when the current Spec anyhow overrides that
API. (Normally, import of duplicate API is not allowed).

* Fix linter errors introduced by PR #279