CNS-315: support subscription renewal
Subscription renewal happens when buying a subscription for a consumer with an
existing subscription, with the same plan (index and version). The existing
subscription is extended by the new duration requested. The total duration is
limited to 1 year, but also 1 additional month to allow renewal with yearly
discount before an existing subscription ends.