CNS-311 fixation store prefix bug (#339)
* CNS-311: (FixationStore) sanitize indices of stored entries

The FixationStore assumes that entries cannot have index's that are prefix of
other entries' indices. However, this was not enforced, so for example: a user
may create a project called "project" and another called "project2", breaking
the logic that relies on unique prefixes.

To fix, require entry names (index) to contain only visible ascii characters
(ascii values 32-125), and use the ascii 'DEL' (invisible) character internally
to terminate the index values when stored. This ensures that no two indices can
ever overlap.

This properly supports Bech32 addresses, which are limited to use only visible
ascii characters as per https://en.bitcoin.it/wiki/BIP_0173#Specification.

* CNS-311: (FixationStore) GetEntry() to always get the latest version

GetEntry() essentially takes a snapshot of an latest version of an entry. The
caller may not choose an earlier version to snapsho - so drop the block arg
from its definition.

* CNS-311: (FixationStore) FindEntry to skip stale entries

The garbage collection mechanism may leave entries with refcount zero for
extended periods. However, after the STALE_ENTRY_TIME they should no longer be
visible via FindEntry() calls - fix the function to enforce this behavior.

* CNS-311: (FixationStore) code clean up

Remove unneeded functions, improve readability, fix comments.

* CNS-311: (FixationStore) refactor code

* CNS-311: (FixationStore) test staleness since when recount reaches zero

Change the test whether an entry is stale (and needs to be deleted or ignored)
to check the time since the entry's refcount becomes zero (and not since the
entry's creation time).

* CNS-311: (FixationStore) standardize unit tests with templates

* CNS-311: fix Plans tests of stale plans

* CNS-311: (review) add sanity test for safeIndex (and minor refactor)