upgrades: add upgrade to v0.8.1
In v0.8.1 the StoreKey of module Projects changed from "projects" to "project".
The store upgrade handler deletes the old storage and adds the new one.

(Note that as of cosmos-sdk 0.45.11 the "Renamed" method for store upgrade did
not work properly).

Disable the attempt for v0.8.0 (which left out the above change), because it
led to a corrupt store state, and could not be easily re-upgraded to a working
state.