CNS-337: sort out test timeout
The e2e timeout was inconsistent:

First, "go test" default timeout is 10 minutes, which it can be changed using
"go test -timeout ...". However, the e2e test internally runs another "go test"
without any argument, so even if the user specified a longer timeout for the
test, the latter would fail within the default timeout.

Second, in at least one place the timeout was hard-coded fixed to 10 minutes.

This is fixed by settings the timeout the intenral "go test" to zero, and then
insteead settings the context.WithTimeout to the timeout requested by the user.