fixation: improve API for GetEntry,FindEntry,PutEntry
Change GetEntry,FindEntry to not return an error, use MustUnmarshal() instead
of Unmarshal(). Change PutEntry to not return anything and not fill the entry
whose refcount is being dropped (caller must already have it).

Previously they returned an error if the Ummarshal of the entry data failed.
However that was bad for two reasons: first, such a failure would indicate a
severe logical bug because the very same data must have been marshaled and
saved before - hence it must be either corrupted or a new lava version now
expects some different format (either way it's a bug). Second, callers do not
have a way to rectify such errors, they just treat it as "not found", making it
useless.

Signed-off-by: Oren Laadan <orenl@lavanet.xyz>