CNS-385: methods GetEntryVersionRange()
Add method GetEntryVersionRange(ctx, index, block, ceiling), which returns
list of block versions of an entry, starting from the nearest-smaller block
and capped by ceiling.

If block is 0 and ceiling is math.MaxUint64, then this gives all versions.
If ceiling is block + STALE_ENTRY_TIME then this returns range of versions
that are may belong in the same "snapshot" (e.g. for a project).