CNS-374: fix vrfIndex mismatch between consumer and provider
Following a data reliability message, the provider verifies that it indeed is
the designated provider by checking that its "self" vrfIndex matches that of
the client.

This got broken with the use of subscription in E2E tests, becuase the default
plan uses 5 providers while E2E uses 2 providers, the while the consumer uses
the number of providers from the pairing, the provider gets it from the plan.
Since they disagree, the VRF output also (likely) differs and triggers error.

The error message was:

    testutil/e2e/logs/05_LavaProvider_01_errors.log:Apr  2 08:10:13 ERR
    Provider identified invalid vrfIndex in data reliability request, the given
    index and self index are different GUID=16864668971218711883....

Actually, in the non-subscription case, the providers gets that number from
the params. Therefore, similar issue would happen also in any case where the
number of providers in a consumer's current pairing differs from that saved in
the params.

This commit fixes the issue by teaching the provider to use the pairing to get
the respective data. It also adds it to `ProviderSessionsWithConsumer struct`
for fast access.