CNS-337: reduce test time by removing redundant client tests
Initially subscription testing was added on top of staked clients, so it
doubled the (already long) test time.

However, it's enough to test staked-clients on ETH1 and subscriptions on
LAV1, as it still provides full coverage of the core functionalty (while
each tests different access methods).

Also generalize the code that invokes client tests so that it can handle
multiple clients (currently up to 3, although we use only 1).