CNS-420: fixation to maintain DeleteAt on future entries too
Previously, entry.DeleteAt was tracked always on the latest entry, even if
there is a future entry with smaller block then this DeleteAt. Such case
prevented us from considering the this DeleteAt in lookup for future entries.

With this fix, entry.DeleteAt will now always be set on the entry with the
largets block that is still before the DeleteAt block (either the latest one
or a future one).