CNS-457: plug the new module into app/app.go and testutils
The purpose of the new protocol module is to hold the parameters for the
protocol versions, i.e. provider and consumer version (and min-version),
on-chain. The can be modified using standard param-change proposals.

Because of limitations of the x/params support for parameters, validation
of new parameters is limited: it cannot verify that provider and consumer
versions are in sync (which is a temp requirement), and it cannot verify
that a min-version is not greater than the corresponding version.