CNS-487: add tests for the case of double-nested deletion
Subscriptions can add/delete projects; Projects can add/delete keys;

Double-nested delete happens the user deletes a project and then in the same
epoch also deletes a key (and also the reverse: first key then project, in
the same epoch).

Double-nested delete also happens when the subscription expires in the same
epoch that a user deleted a project (of that subscription).

Add tests that verify that these events are handled correctly without errors
or a panic (specifically, due to stray timer).