CNS-490: handle putEntry() of a future entry
Future entries always carry a refcount 1 and cannot be references.
If they are deleted, e.g. by explicit PutEntry() or by an implicit
putEntry from trimming of future entries - they nneed not undergo
stale-period like regular entries.

Also, move the trimming of future entries upon DelEntry() - when
marking an entry with DeleteAt (instead of when deletion actually
is due, which can be later). This, together with AppendEntry which
disallows new entries beyond a DeleteAt mark - ensures that there
are never extranneous entries beyond the delete point.