CNS-489: initialize version(s) of fixation-store, timer-store
The versions of FixationStore and TimerStore were not initialized properly when
starting a new network.

Fix by calling setVersion() at the beginning of the Init() (from genesis), to
force-set the current version. If the genesis holds previous state data, then
that data -with the saved version- will override.

(Also re-arrange the code to alwways place Initxxx() and Exportxxx() functions
in x/xxx/keeper/keeper.go)