CNS-489: rename timer's version key to "Timer_Version"
The previous name was not only inaccurate but also identical to that of the
fixation store version; And since the fixation store uses the timer store such
that they share the same store namespace (prefix), this created a conflict.

To fix this we change the timer-store version key suitably to "Timer_Version".
Doing standard migration is tricky because it needs to use the previous key to
get the previous version and the new key to set the new version.

Instead, because there is no change in the underlying store, it is sufficient
to simply write the new (i.e. current) version using the new key. This is done
automatically in the timer-store's Init() function.

Note that we can't simply remove the old key to cleanup, since in the context
of the fixation store it will delete the latter's version. We need migration
for this (that would only delete for timer-store that do not belong to some
fixation-store). This cleanup is deferred to a next version.