CNS-516: introduce utils/rand for deterministic PRNG
Using the global rand directly is not safe in terms of determinism, e.g. if
another go routine calls (now or in the future) it to get a random number at
the wrong time. Likewise, rand.Seed() is deprecated in Go (for very similar
reasons).

Also, Go does not guarantee that the PRNG driving the default source for rand
will always provide the same output for a given seed - across different Go
versions.

Instead, implement our own source (PRNG) that combines rand.New() to provide
a dedicated (private) instance of a PRNG that uses the xoroshiro256** PRNG
from the the xorshift package as a source.

(See https://en.wikipedia.org/wiki/Xorshift for more details)