CNS-597: fixation: fix DelEntry() of future entry and no previous entry
DelEntry() of future entry normally searches for a strictly nearest-smaller
previous entry (rather than nearest-no-later). For rationale see comments in
the code.

However, that logic assumes that there is at least one previous version to
the affected future entry, and otherwise fails to delete a future entry that
does not have such previous current version.

To fix that, check for this kind of exception - a (first) future entry that
has not matured yet, i.e. that has no current version of the entry in place.
If found, then trim the entire cadence.

Also fix a missing call to delete the timer of the canceled future entry in
cancelEntry(); Rename to putFutureEntry() and refactor trimFutureEntries().

Add unit test for this DelEntry() special case.