feat: add slices helper Filter()
Filter() takes a slice and returns a sub-slice based on an input filter
function applied on each of the elements of the slice.

Also rename the old Filter() to what it really is - Map(), which take a
slice and returns a new slice with the result of applying the input map
function on the elements of input slice.